refactor: move code to a single file
